delete Listing

+++ api/controllers/listingController.js
@@ -35,6 +35,8 @@ export const getUserListing = async (req , res , next)=>{
 
 export const deleteListing = async (req , res , next)=>{
 
+        const listing=await Listing.findById(req.params.id);
+        if(!listing) return next(errorHandler(404 , "Listing not found!"))
         try {
             await Listing.findByIdAndDelete(req.params.id);
             res.status(200).json('Listing has been deleted!');
